requirements.txt
Flask
flask-wtf
PyMuPDF
reportlab
python-docx
matplotlib
requests
pytube
google-api-python-client
yt-dlp
